# user_service/pkg/api/swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "User API",
    "version": "1.0.0",
    "contact": {
      "name": "Dokhoyan dev.",
      "email": "dokhoyanjobb@mail.ru"
    }
  },
  "tags": [
    {
      "name": "UserV1"
    }
  ],
  "host": "localhost:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user/v1": {
      "delete": {
        "summary": "Удаляет пользователя по id",
        "operationId": "UserV1_Delete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "Идентификатор удаляемого пользователя",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "UserV1"
        ]
      },
      "patch": {
        "summary": "Обновляет пользователя по id",
        "operationId": "UserV1_Update",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user_v1UpdateRequest"
            }
          }
        ],
        "tags": [
          "UserV1"
        ]
      }
    },
    "/user/v1/auth-data/{username}": {
      "get": {
        "operationId": "UserV1_GetUserAuthData",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/user_v1GetUserAuthDataResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "UserV1"
        ]
      }
    },
    "/user/v1/create": {
      "post": {
        "summary": "Создает нового пользователя в системе",
        "operationId": "UserV1_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/user_v1CreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user_v1CreateRequest"
            }
          }
        ],
        "tags": [
          "UserV1"
        ]
      }
    },
    "/user/v1/{id}": {
      "get": {
        "summary": "Возвращает пользователя по id",
        "operationId": "UserV1_Get",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/user_v1GetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "UserV1"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "user_v1CreateRequest": {
      "type": "object",
      "properties": {
        "info": {
          "$ref": "#/definitions/user_v1UserInfo"
        },
        "pass": {
          "$ref": "#/definitions/user_v1UserPassword"
        }
      }
    },
    "user_v1CreateResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64",
          "title": "Идентификатор созданного пользователя"
        }
      }
    },
    "user_v1GetResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/user_v1User"
        }
      }
    },
    "user_v1GetUserAuthDataResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/user_v1UserAuthData"
        }
      }
    },
    "user_v1UpdateInfo": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "Новое имя пользователя"
        },
        "username": {
          "type": "string",
          "title": "Имя учетной записи"
        },
        "email": {
          "type": "string",
          "title": "Новая почта пользователя"
        },
        "birthDate": {
          "type": "string",
          "format": "date-time",
          "title": "Дата рождения пользователя"
        },
        "avatarUrl": {
          "type": "string",
          "title": "Фото профиля"
        },
        "role": {
          "$ref": "#/definitions/user_v1UserRole"
        }
      }
    },
    "user_v1UpdateRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64",
          "title": "Идентификатор обновляемого пользователя"
        },
        "info": {
          "$ref": "#/definitions/user_v1UpdateInfo"
        }
      }
    },
    "user_v1User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64",
          "title": "Идентификатор пользователя"
        },
        "info": {
          "$ref": "#/definitions/user_v1UserInfo",
          "title": "Данные пользователя"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "title": "Дата создания пользователя"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "title": "Дата обновления пользователя"
        }
      }
    },
    "user_v1UserAuthData": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "username": {
          "type": "string"
        },
        "passwordHash": {
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/user_v1UserRole"
        }
      }
    },
    "user_v1UserInfo": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "Имя пользователя"
        },
        "username": {
          "type": "string",
          "title": "Имя учетной записи"
        },
        "email": {
          "type": "string",
          "title": "Почта пользователя"
        },
        "birthDate": {
          "type": "string",
          "format": "date-time",
          "title": "Дата рождения пользователя"
        },
        "avatarUrl": {
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/user_v1UserRole"
        }
      }
    },
    "user_v1UserPassword": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string",
          "title": "Пароль пользователя"
        },
        "passwordConfirm": {
          "type": "string",
          "title": "Подтверждение пароля пользователя"
        }
      }
    },
    "user_v1UserRole": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "USER",
        "ADMIN"
      ],
      "default": "UNKNOWN",
      "title": "- UNKNOWN: Не определён\n - USER: Простой пользователь\n - ADMIN: Администратор"
    }
  }
}